PLATUI-64: Sanitised Json output. Fixed inCompleteAlerts for Axe and tested with a sample incomplete Alerts. Refactor.
