+++ parse-a11y-reports/src/main/scala/uk/gov/hmrc/a11y/AxeReport.scala
@@ -1,9 +1,8 @@
 package uk.gov.hmrc.a11y
 
-import java.io.FileInputStream
-
-import play.api.libs.json.{JsValue, Json, OWrites}
-import uk.gov.hmrc.a11y.ParseA11yReport.{outputFileWriter, testSuite}
+import play.api.libs.json.{JsString, JsValue, Json}
+import uk.gov.hmrc.a11y.JsonUtil._
+import uk.gov.hmrc.a11y.ParseA11yReport.testSuite
 
 object AxeReport {
 
@@ -33,38 +32,13 @@ object AxeReport {
       t =>
         val code = getValue(t, "id")
         val severity = getValue(t, "impact")
-        val description = getValue(t, "message")
-        val selector = getValue(t, "data")
-        val snippet = getValue(t, "html")
-        Violation("axe", testSuite, pageUrl, testRunTimeStamp, timeStamp, code, severity, s"Incomplete Alert: $description", selector, snippet)
+        val description = getValue(t, "description")
+        val nodes = (t \ "nodes").as[List[JsValue]]
+        val selector = getValue(nodes.head, "target")
+        val snippet = getValue(nodes.head, "html")
+        Violation("axe", testSuite, pageUrl, testRunTimeStamp, timeStamp, code, severity, JsString(s"Incomplete Alert: $description"), selector, snippet)
     }
 
-    val alerts = violationAlerts ++ inCompleteAlerts
-
-    implicit val reportWrites: OWrites[Violation] = Json.writes[Violation]
-    alerts.foreach { v =>
-      println(s"""{"index":{"_index":"accessibility","_type":"alerts"}}\n${Json.toJson(v).toString()}\n""")
-      outputFileWriter.write(s"""{"index":{"_index":"accessibility","_type":"alerts"}}\n${Json.toJson(v).toString()}\n""")
-    }
-
+    Output.writeOutput(violationAlerts ++ inCompleteAlerts)
   }
-
-  def parseJsonFile(reportPath: String): JsValue = {
-    val stream = new FileInputStream(reportPath)
-    try {
-      Json.parse(stream)
-    }
-    finally {
-      stream.close()
-    }
-  }
-
-  def getValue(json: JsValue, field: String): String = {
-
-    (json \ field).toEither match {
-      case Left(s) => "UNDEFINED"
-      case Right(value) => value.toString()
-    }
-  }
-
 }
