+++ parse-a11y-reports/src/main/scala/uk/gov/hmrc/a11y/ParseA11yReport.scala
@@ -29,7 +29,6 @@ object ParseA11yReport {
     println("********** Generating A11Y report ***********")
     generateReport()
     println("********** Completed ***********")
-
   }
 
   def generateReport(): Unit = {
