+++ parse-a11y-reports/src/main/scala/uk/gov/hmrc/a11y/Violation.scala
@@ -1,4 +1,6 @@
 package uk.gov.hmrc.a11y
 
-case class Violation(tool: String, testSuite: String, path: String, testRun: String, timeStamp: Long, code: String, severity: String,
-                     description: String, selector: String, snippet: String)
+import play.api.libs.json.JsValue
+
+case class Violation(tool: String, testSuite: String, path: String, testRun: String, timeStamp: Long, code: JsValue, severity: JsValue,
+                     description: JsValue, selector: JsValue, snippet: JsValue)
