+++ parse-a11y-reports/src/main/scala/uk/gov/hmrc/a11y/VnuReport.scala
@@ -2,9 +2,9 @@ package uk.gov.hmrc.a11y
 
 import java.io.File
 
-import play.api.libs.json.{JsValue, Json, OWrites}
-import uk.gov.hmrc.a11y.AxeReport.{getValue, parseJsonFile}
-import uk.gov.hmrc.a11y.ParseA11yReport.{outputFileWriter, testSuite}
+import play.api.libs.json.{JsValue, Json}
+import uk.gov.hmrc.a11y.JsonUtil._
+import uk.gov.hmrc.a11y.ParseA11yReport.testSuite
 
 object VnuReport {
 
@@ -16,7 +16,7 @@ object VnuReport {
     new File(vnuReport).length() match {
       case 0 => println(s"vnu report is empty for $vnuReport")
       case _ =>
-        val alerts = (parseJsonFile(vnuReport) \ "messages").as[List[JsValue]].map{
+        val alerts = (parseJsonFile(vnuReport) \ "messages").as[List[JsValue]].map {
           t =>
             val code = getValue(t, "code")
             val severity = getValue(t, "type")
@@ -26,11 +26,7 @@ object VnuReport {
 
             Violation("vnu", testSuite, pageUrl, testRunTimeStamp, timeStamp, code, severity, description, selector, snippet)
         }
-        implicit val reportWrites: OWrites[Violation] = Json.writes[Violation]
-        alerts.foreach { v =>
-          println(s"""{"index":{"_index":"accessibility","_type":"alerts"}}\n${Json.toJson(v).toString()}\n""")
-          outputFileWriter.write(s"""{"index":{"_index":"accessibility","_type":"alerts"}}\n${Json.toJson(v).toString()}\n""")
-        }
+        Output.writeOutput(alerts)
     }
 
   }
